Implement "fixed" interaction + other utilities
The interaction NetworkInteraction allows for arbitrary interactions
defined through a network (a list of which agents interacts with which).
 This is particularly useful for simulations where the network is fixed
 because it is much faster than checking for neighbors every time.

Community includes several new methods useful to simulate fixed,
regular arrays of agents:
    1. regular_positions(): sets the agents in a regular N-dimensional
    lattice.
    2. build_network(nneis, network): defines the "interaction network"
    (i.e. a list of lists with each agent's neighbors). This can be fed
    into a NetworkInteraction in order to "freeze" the interaction.

An example of how to compute the susceptibility in a regular lattice is
included in examples/susceptibility/vicsek_regular.

Everything related to the generation of random numbers has been moved
to random.cpp and uses the efficient Ziggurat method now.